+++ src/file_management/node_entry_meta.rs
@@ -1,6 +1,3 @@
-//! TODO: Remove if used
-#![allow(dead_code)]
-
 use crate::prelude::*;
 use std::path::PathBuf;